Refactor SinBin and StableDiffusion commands to classes
Migrated SinBin and StableDiffusion commands from function-based implementations in OldCommands to class-based implementations in the Commands directory, inheriting from AbstractCommand. Updated BotUtils to support new command structure and improved isAdmin logic. Removed old command files and cleaned up related code in CommandHandler.
